requirements.txt
# Agent Framework
langgraph==1.1.3
langchain==1.2.13
langchain-core==1.2.20
langchain-openai==1.1.11

# Vector DB
qdrant-client==1.17.1
langchain-qdrant==1.1.0

# Data Pipeline (문서 파싱)
pymupdf==1.27.2.2

# API Server
fastapi==0.119.0
uvicorn[standard]==0.37.0
pydantic==2.12.0

# Observability / Eval
langfuse==4.0.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.0